Updated store4 to v4.0.2 to fix problems with date and incompatibility with kotlin 1.5 in tests.
MAILAND-1875

+++ app/src/main/java/ch/protonmail/android/mapper/bridge/AddressKeysBridgeMapper.kt
@@ -59,7 +59,7 @@ class AddressKeyBridgeMapper @Inject constructor() : BridgeMapper<OldKey, Addres
 
     override fun OldKey.toNewModel() = AddressKey(
         id = Id(id),
-        version = 4u, // TODO not implemented on old Keys
+        version = 4.toUInt(), // TODO not implemented on old Keys
         canEncrypt = canEncrypt(flags),
         canVerifySignature = canVerifySignature(flags),
         publicKey = PgpField.PublicKey(NotBlankString("none")), // TODO not implemented on old Keys!!!
