+++ app/src/main/java/ch/protonmail/android/mapper/bridge/UserKeysBridgeMapper.kt
@@ -59,7 +59,7 @@ class UserKeyBridgeMapper @Inject constructor() : BridgeMapper<OldKey, UserKey> 
 
     override fun OldKey.toNewModel() = UserKey(
         id = Id(id),
-        version = 4u, // TODO not implemented on old Keys
+        version = 4.toUInt(), // TODO not implemented on old Keys
         privateKey = PgpField.PrivateKey(NotBlankString(privateKey)),
         token = getToken(token)
     )
